🚚 move election round selector where it belongs
