+++ packages/justiciar/src/voter.ts
@@ -2,11 +2,9 @@ import type { CtorToolkit, MoleculeType } from "atom.io"
 import { moleculeFamily, selectorFamily } from "atom.io"
 import { findRelations } from "atom.io/data"
 
-import { candidateStatusSelectors, electionRoundCandidateMolecules } from "./candidate"
-import { electionMolecules, votes } from "./election"
+import { candidateStatusSelectors } from "./candidate"
+import { votes } from "./election"
 import type { ElectionRoundKey } from "./election-round"
-import { electionRoundMolecules } from "./election-round"
-import { Rational } from "./rational"
 
 export type Voter = {
 	type: `voter`
@@ -89,53 +87,3 @@ export const electionRoundVoterMolecules = moleculeFamily({
 export type ElectionRoundVoterInstance = InstanceType<
 	MoleculeType<typeof electionRoundVoterMolecules>
 >
-
-export type ElectionRoundVoteTotal = [candidateId: string, voteTotal: Rational]
-export const electionRoundVoteTotalsSelectors = selectorFamily<
-	ElectionRoundVoteTotal[],
-	ElectionRoundKey
->({
-	key: `electionRoundVoteTotals`,
-	get:
-		(keys) =>
-		({ get }) => {
-			const election = get(electionMolecules, keys.election)
-			const electionRound = get(electionRoundMolecules, keys)
-
-			const candidates = get(election.state.candidates.relatedKeys)
-			const runningCandidates = candidates.filter((candidateKey) => {
-				const candidateRoundKey = {
-					electionRound: keys,
-					candidate: candidateKey,
-				}
-				const candidate = get(electionRoundCandidateMolecules, candidateRoundKey)
-				const status = get(candidate.state.status)
-				return status === `running`
-			})
-
-			const voteTotals = runningCandidates
-				.map<ElectionRoundVoteTotal>((candidateKey) => {
-					const candidateTotalVote = new Rational()
-					const candidateVoterKeys = get(findRelations(votes, candidateKey).voterKeysOfCandidate)
-
-					const myVoteDenominators = candidateVoterKeys
-						.map<bigint | null>((voterKey) => {
-							const voter = electionRound.voters.get(voterKey) ?? electionRound.spawnVoter(voterKey)
-							const [voterTopFavorites] = get(voter.state.favorites)
-							if (!voterTopFavorites.includes(candidateKey)) {
-								return null
-							}
-							const denominator = BigInt(voterTopFavorites.length)
-							return denominator
-						})
-						.filter((total) => total !== null)
-
-					for (const denominator of myVoteDenominators) {
-						candidateTotalVote.add(1n, denominator)
-					}
-					return [candidateKey, candidateTotalVote]
-				})
-				.sort(([, totalA], [, totalB]) => (totalA.isGreaterThan(totalB) ? -1 : 1))
-			return voteTotals
-		},
-})
